feat(authz): added type guards and moved types
