+++ src/lib/interactionManager/types.ts
@@ -29,6 +29,27 @@ export enum FlowType {
   Authorization = 'Authorization',
 }
 
+export enum AuthorizationType {
+  AuthorizationRequest = 'authorizationRequest',
+  AuthorizationResponse = 'authorizationResponse',
+}
+
+export interface AuthorizationResponse {
+  description: string
+  imageURL?: string
+  action?: string
+}
+
+export interface AuthorizationRequest extends AuthorizationResponse {
+  callbackURL: string
+}
+
+export interface AuthorizationFlowState {
+  description: string
+  imageURL?: string
+  action?: string
+}
+
 export interface AuthenticationFlowState extends FlowState {
   description: string
 }
